+++ packages/web-components/keen-slider-navigation-dots/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/web-component-keen-slider-navigation-dots
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/web-component-keen-slider@1.0.14
+
 ## 1.0.12
 
 ### Patch Changes