+++ packages/web-components/keen-slider-scrollbar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/web-component-keen-slider-scrollbar
 
+## 1.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/web-component-keen-slider@1.0.14
+
 ## 1.0.19
 
 ### Patch Changes